Refactor location of push logic to enable code sharing
